# question-engine/openapi/question-engine.v1.yaml
openapi: 3.0.3
info:
  title: Question Engine API
  version: 1.0.0
  description: >
    question-engine 本期稳定能力契约。该契约面向公司教育生态平台，
    覆盖能力目录、题目加工任务、标准题目包、OCR/AI/导出/file/callback
    运行时入口；本地小平台页面兼容 API 不属于本契约主路径。
servers:
  - url: http://localhost:8018
    description: Local Java backend
...
security:
  - PlatformBearerAuth: []
    TenantHeader: []
    OperatorHeader: []
paths:
...
  /api/import-tasks/{jobId}/image-library:
    get:
      tags: [review-workbench, file-flow]
      operationId: getImportTaskImageLibrary
      summary: 获取本次导入任务生成的题图库
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Import task image library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionImageLibrary"
  /api/import-tasks/{jobId}/questions/{questionId}/images:
    post:
      tags: [review-workbench, file-flow]
      operationId: uploadImportQuestionImages
      summary: 给导入题本地上传题图
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/QuestionId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Updated question images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionImageMutationResult"
  /api/import-tasks/{jobId}/questions/{questionId}/images/select:
    post:
      tags: [review-workbench, file-flow]
      operationId: selectImportQuestionImages
      summary: 从任务题图库选择图片挂到当前题
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/QuestionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectQuestionImagesRequest"
      responses:
        "200":
          description: Updated question images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionImageMutationResult"
  /api/import-tasks/{jobId}/questions/{questionId}/images/{imageId}:
    get:
      tags: [review-workbench, file-flow]
      operationId: getImportQuestionImage
      summary: 读取导入题题图文件
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/QuestionId"
        - name: imageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image binary
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
...
  /api/question-bank/questions/{questionId}/image-library:
    get:
      tags: [review-workbench, file-flow]
      operationId: getBankQuestionImageLibrary
      summary: 获取题库题可复用题图库
      parameters:
        - $ref: "#/components/parameters/QuestionId"
      responses:
        "200":
          description: Bank question image library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionImageLibrary"
  /api/question-bank/questions/{questionId}/images:
    post:
      tags: [review-workbench, file-flow]
      operationId: uploadBankQuestionImages
      summary: 给题库题本地上传题图
      parameters:
        - $ref: "#/components/parameters/QuestionId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Updated bank question images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionImageMutationResult"
  /api/question-bank/questions/{questionId}/images/{imageId}:
    get:
      tags: [review-workbench, file-flow]
      operationId: getBankQuestionImage
      summary: 读取题库题题图文件
      parameters:
        - $ref: "#/components/parameters/QuestionId"
        - name: imageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image binary
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    PlatformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 生产环境由平台网关或平台服务注入；本地开发可省略。
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: 平台租户或机构上下文，生产环境必填。
    OperatorHeader:
      type: apiKey
      in: header
      name: X-Operator-Id
      description: 平台操作者或服务账号上下文，生产环境必填。
  parameters:
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
    QuestionId:
      name: questionId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    QuestionImage:
      type: object
      properties:
        id:
          type: string
        imageId:
          type: string
        index:
          type: integer
        imageIndex:
          type: integer
        name:
          type: string
        path:
          type: string
        url:
          type: string
        source:
          type: string
        type:
          type: string
        size:
          type: integer
          format: int64
        storageFileId:
          type: string
        questionId:
          type: string
        raw:
          type: object
          additionalProperties: true
    QuestionImageLibrary:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuestionImage"
    SelectQuestionImagesRequest:
      type: object
      properties:
        imageIds:
          type: array
          items:
            type: string
        images:
          type: array
          description: 可传完整图片对象；服务端会按 imageId/id/url/path 从任务题图库匹配。
          items:
            $ref: "#/components/schemas/QuestionImage"
    QuestionImageMutationResult:
      type: object
      required: [images]
      properties:
        images:
          type: array
          items:
            $ref: "#/components/schemas/QuestionImage"
        uploaded:
          type: array
          items:
            $ref: "#/components/schemas/QuestionImage"
        selected:
          type: array
          items:
            $ref: "#/components/schemas/QuestionImage"
        question:
          type: object
          additionalProperties: true
        task:
          type: object
          additionalProperties: true
...